@@ -6,62 +6,69 @@
 		"dev": "bun --bun vite dev",
 		"build": "bun --bun vite build",
 		"preview": "vite preview",
-		"test": "npm run test:integration && npm run test:unit",
 		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
-		"test:integration": "playwright test",
-		"test:unit": "vitest",
 		"lint": "prettier --check .",
 		"format": "prettier --write .",
 		"db:generate": "drizzle-kit generate",
 		"db:push": "drizzle-kit push"
 	},
 	"devDependencies": {
-		"@playwright/test": "^1.52.0",
-		"@skeletonlabs/skeleton": "^2.10.4",
-		"@skeletonlabs/tw-plugin": "^0.4.0",
+		"@internationalized/date": "^3.8.2",
+		"@lucide/svelte": "^0.525.0",
 		"@sveltejs/adapter-auto": "^6.0.1",
-		"@sveltejs/kit": "^2.21.0",
-		"@sveltejs/vite-plugin-svelte": "^5.0.3",
+		"@sveltejs/kit": "^2.22.4",
+		"@sveltejs/vite-plugin-svelte": "^5.1.0",
 		"@tailwindcss/forms": "^0.5.10",
-		"@types/bun": "^1.2.13",
-		"@types/node": "^22.15.18",
+		"@tailwindcss/typography": "^0.5.16",
+		"@tailwindcss/vite": "^4.1.11",
+		"@types/bun": "^1.2.18",
+		"@types/js-cookie": "^3.0.6",
+		"@types/node": "^24.0.12",
 		"@types/uuid": "^10.0.0",
 		"autoprefixer": "^10.4.21",
-		"drizzle-kit": "^0.31.1",
-		"postcss": "^8.5.3",
-		"prettier": "^3.5.3",
+		"bits-ui": "^2.8.10",
+		"clsx": "^2.1.1",
+		"drizzle-kit": "^0.31.4",
+		"formsnap": "^2.0.1",
+		"mode-watcher": "^1.1.0",
+		"postcss": "^8.5.6",
+		"prettier": "^3.6.2",
 		"prettier-plugin-svelte": "^3.4.0",
-		"svelte": "5.30.1",
-		"svelte-check": "^4.2.1",
-		"tailwindcss": "^3.4.17",
+		"prettier-plugin-tailwindcss": "^0.6.14",
+		"shiki": "^3.7.0",
+		"svelte": "5.35.5",
+		"svelte-check": "^4.2.2",
+		"svelte-sonner": "^1.0.5",
+		"tailwind-merge": "^3.3.1",
+		"tailwind-variants": "^1.0.0",
+		"tailwindcss": "^4.1.11",
 		"tslib": "^2.8.1",
+		"tw-animate-css": "^1.3.5",
 		"typescript": "^5.8.3",
+		"vaul-svelte": "^1.0.0-next.7",
 		"vite": "^6.3.5",
-		"vite-plugin-wasm": "^3.4.1",
-		"vitest": "^3.1.3"
+		"vite-plugin-wasm": "^3.5.0"
 	},
 	"type": "module",
 	"dependencies": {
-		"@felte/validator-zod": "^1.0.18",
-		"@mistralai/mistralai": "^1.6.0",
-		"@neondatabase/serverless": "^1.0.0",
-		"@sveltejs/adapter-node": "^5.2.12",
-		"@sveltejs/enhanced-img": "^0.6.0",
-		"devalue": "^5.1.1",
+		"@mistralai/mistralai": "^1.7.3",
+		"@neondatabase/serverless": "^1.0.1",
+		"@sveltejs/adapter-node": "^5.2.13",
+		"@sveltejs/enhanced-img": "^0.6.1",
 		"dexie": "^4.0.11",
 		"dotenv": "^16.5.0",
-		"drizzle-orm": "^0.43.1",
-		"felte": "^1.3.0",
+		"drizzle-orm": "^0.44.2",
 		"highlight.js": "^11.11.1",
-		"lucide-svelte": "^0.511.0",
-		"marked": "^15.0.11",
-		"marked-katex-extension": "^5.1.4",
+		"js-cookie": "^3.0.5",
+		"marked": "^16.0.0",
+		"marked-katex-extension": "^5.1.5",
 		"pretty-bytes": "^7.0.0",
+		"svelte-persisted-store": "^0.12.0",
 		"tiktoken": "^1.0.21",
 		"ufo": "^1.6.1",
 		"uuid": "^11.1.0",
-		"zod": "^3.24.4"
+		"zod": "^4.0.2"
 	},
 	"trustedDependencies": [
 		"@sveltejs/kit",
