@@ -1,14 +1,17 @@
 <script lang="ts">
-	import { fade } from 'svelte/transition';
 	import { goto } from '$app/navigation';
-	import TriangleAlertIcon from 'lucide-svelte/icons/triangle-alert';
-	import LogOutIcon from 'lucide-svelte/icons/log-out';
-	import Repeat2Icon from 'lucide-svelte/icons/repeat-2';
+	import TriangleAlertIcon from '@lucide/svelte/icons/triangle-alert';
+	import LogOutIcon from '@lucide/svelte/icons/log-out';
+	import Repeat2Icon from '@lucide/svelte/icons/repeat-2';
 	import { loadModels, models } from '$lib/stores/models.svelte';
 	import { apiKey } from '$lib/stores/apiKey';
+	import Cookies from 'js-cookie';
+	import * as Alert from '$lib/components/ui/alert/index.js';
+	import Button from '$lib/components/ui/button/button.svelte';
 
 	function deleteApiKey() {
 		apiKey.set('');
+		Cookies.set('apiKey', '');
 		models.error = null;
 		goto('/');
 	}
@@ -19,38 +22,21 @@
 </script>
 
 {#if models.error}
-	<div class="alert variant-ghost-error">
-		<div>
-			<TriangleAlertIcon size={24} />
-		</div>
-		<div class="alert-message">
-			<h3 class="text-xl">{models.error.title}</h3>
+	<Alert.Root variant="destructive">
+		<TriangleAlertIcon />
+		<Alert.Title>{models.error.title}</Alert.Title>
+		<Alert.Description>
 			<p>{models.error.message}</p>
-			<div class="flex flex-row gap-2 items-center">
-				<button
-					type="button"
-					class="btn justify-start font-bold transition-all variant-ringed-primary disabled:opacity-75"
-					transition:fade
-					disabled={models.loading}
-					onclick={deleteApiKey}
-				>
-					<LogOutIcon class="flex-shrink-0" />
+			<div class="flex flex-row items-center gap-2">
+				<Button variant="secondary" disabled={models.loading} onclick={deleteApiKey}>
+					<LogOutIcon class="shrink-0" />
 					<span class="truncate">Delete API key</span>
-				</button>
-				<button
-					type="button"
-					class="btn justify-start font-bold transition-all variant-ringed-secondary disabled:opacity-75"
-					transition:fade
-					disabled={models.loading}
-					onclick={reloadModels}
-				>
-					<Repeat2Icon class="flex-shrink-0" />
+				</Button>
+				<Button variant="destructive" disabled={models.loading} onclick={reloadModels}>
+					<Repeat2Icon class="shrink-0" />
 					<span class="truncate">Retry</span>
-				</button>
+				</Button>
 			</div>
-		</div>
-	</div>
+		</Alert.Description>
+	</Alert.Root>
 {/if}
-
-<style>
-</style>
