@@ -1,7 +1,7 @@
 import { Mistral } from '@mistralai/mistralai';
 import { normalizeURL } from 'ufo';
 
-const defaultEndpoint = 'https://api.mistral.ai';
+export const defaultEndpoint = 'https://api.mistral.ai';
 
 export function getClientForRequest(params: { apiKey: string; endpoint?: string }) {
 	const useEndpoint = params.endpoint?.length ? params.endpoint : defaultEndpoint;
