@@ -1,13 +1,20 @@
 <script lang="ts">
 </script>
 
-<div class="flex flex-grow flex-shrink items-stretch flex-col gap-4 p-4 max-h-[calc(100vh-88px)] lg:max-h-screen">
-	<h1 class="text-3xl">About</h1>
-	<div class="card p-4 variant-ghost-primary text-primary-300">
-		<p><strong>Mistral playground</strong> is not affiliated with <strong>MistralAI</strong>.</p>
-	</div>
-	<div class="card p-4 variant-ghost-primary text-primary-300">
-		<p>No data is shared with any other third party other than <strong>Mistral</strong> itself.</p>
-		<p>Your API key and history are only stored locally in your browser.</p>
-	</div>
+<div class="flex max-h-[calc(100vh-88px)] shrink grow flex-col items-stretch gap-4 lg:max-h-screen">
+	<h1 class="scroll-m-20 text-4xl font-extrabold tracking-tight text-balance">About</h1>
+	<p class="leading-7">
+		<strong>Mistral playground</strong> is not affiliated with <strong>MistralAI</strong>.
+	</p>
+	<p class="leading-7">
+		No data is shared with any other third party other than <strong>Mistral</strong> itself.
+	</p>
+	<p class="leading-7">Your API key and history are only stored locally in your browser.</p>
+	<h1 class="scroll-m-20 text-3xl font-extrabold tracking-tight text-balance">Repository</h1>
+	<p class="leading-7">
+		You can find the project code on github : <a
+			href="https://github.com/Glagan/mistral-playground"
+			class="text-primary font-medium underline underline-offset-4">https://github.com/Glagan/mistral-playground</a
+		>.
+	</p>
 </div>
