@@ -1,47 +1,41 @@
 <script lang="ts">
 	import { fade, slide } from 'svelte/transition';
-	import {
-		SlideToggle,
-		focusTrap,
-		type ModalComponent,
-		type ModalSettings,
-		getModalStore
-	} from '@skeletonlabs/skeleton';
 	import { get_encoding } from 'tiktoken';
 	import { apiKey } from '$lib/stores/apiKey';
 	import type { Usage, Message, AssistantMessage, MessageRole, MessageContent } from '$lib/types';
 	import { goto } from '$app/navigation';
 	import { browser } from '$app/environment';
-	import Messages from '$lib/components/Messages.svelte';
+	import Messages from '$lib/components/Chat/Messages.svelte';
 	import { settings } from '$lib/stores/settings';
 	import { onDestroy, onMount } from 'svelte';
 	import { chat } from '$lib/stores/chat.svelte';
 	import { v7 as uuid } from 'uuid';
-	import Settings2Icon from 'lucide-svelte/icons/settings-2';
-	import CircleHelpIcon from 'lucide-svelte/icons/circle-help';
-	import FileTextIcon from 'lucide-svelte/icons/file-text';
-	import ImageUpIcon from 'lucide-svelte/icons/image-up';
+	import FileTextIcon from '@lucide/svelte/icons/file-text';
+	import ImageUpIcon from '@lucide/svelte/icons/image-up';
 	import { loadModels, models } from '$lib/stores/models.svelte';
 	import { getClientForRequest } from '$lib/mistral';
 	import ModelError from '$lib/components/ModelError.svelte';
-	import ShareModal from '$lib/components/ShareModal.svelte';
 	import { defaultChatModel } from '$lib/const';
 	import { fileToB64, handleFileUpload } from '$lib/files';
-	import FileUploadPreview from '$lib/components/FileUploadPreview.svelte';
-	import type { TextChunk } from '@mistralai/mistralai/models/components';
-	import { getToastStore } from '@skeletonlabs/skeleton';
+	import FileUploadPreview from '$lib/components/File/UploadPreview.svelte';
+	import type { ChatCompletionStreamRequest, TextChunk } from '@mistralai/mistralai/models/components';
 	import { editing } from '$lib/stores/editing.svelte';
 	import { db } from '$lib/stores/db';
+	import { toast } from 'svelte-sonner';
+	import Button from '$lib/components/ui/button/button.svelte';
+	import SendHorizontalIcon from '@lucide/svelte/icons/send-horizontal';
+	import { Textarea } from '$lib/components/ui/textarea/index.js';
+	import Badge from '$lib/components/ui/badge/badge.svelte';
+	import Options from './Options.svelte';
+	import * as Drawer from '$lib/components/ui/drawer/index.js';
+	import SlidersHorizontalIcon from '@lucide/svelte/icons/sliders-horizontal';
 
 	if (browser && !$apiKey) {
 		goto('/', { replaceState: true });
 	}
 
 	const encoding = get_encoding('cl100k_base');
-	const modalStore = getModalStore();
-	const toastStore = getToastStore();
 
-	let showOptions = $state(false);
 	let promptText = $state('');
 	let isDragging = $state(false);
 	let dragCounter = $state(0);
@@ -59,7 +53,7 @@
 	const unsubscribe = settings.subscribe((value) => {
 		if (chat.state.messages.length === 0) {
 			chat.state.options.temperature = value.temperature;
-			chat.state.options.randomSeed = value.seed ? Number(value.seed) : undefined;
+			chat.state.options.seed = value.seed ? Number(value.seed) : undefined;
 		}
 	});
 
@@ -76,8 +70,6 @@
 	}
 
 	const tokens = $derived(encoding.encode(promptText).length);
-	let systemPrompt = $state('');
-	const systemPromptTokens = $derived(encoding.encode(systemPrompt).length);
 	const maxTokens = $derived(
 		models.chat.find((model) => model.id === chat.state.options.model)?.maxContextLength ?? 32000
 	);
@@ -112,39 +104,44 @@
 	async function generate(messages: Message[], answer: AssistantMessage) {
 		const outputNode = document.getElementById('messages-container')?.parentElement;
 		loading = true;
-		showOptions = false;
 
 		abortController = new AbortController();
 		const startedAt = performance.now();
 		try {
 			const client = getClientForRequest({ apiKey: $apiKey, endpoint: $settings.endpoint });
+			const finalMessages: ChatCompletionStreamRequest['messages'] = messages.map((message) => {
+				// Useless condition to avoid type errors
+				if (message.role === 'user') {
+					return {
+						role: message.role,
+						content: message.versions[message.index].content
+					};
+				} else if (message.role === 'assistant') {
+					return {
+						role: message.role,
+						content: message.versions[message.index].content
+					};
+				}
+				return {
+					role: message.role,
+					content: message.versions[message.index].content
+				};
+			});
+			if (chat.state.options.systemPrompt) {
+				finalMessages.unshift({ role: 'system', content: chat.state.options.systemPrompt! });
+			}
 			const chatStreamResponse = await client.chat.stream(
 				{
 					model: chat.state.options.model ? chat.state.options.model : defaultChatModel,
-					messages: messages.map((message) => {
-						// Useless condition to avoid type errors
-						if (message.role === 'user') {
-							return {
-								role: message.role,
-								content: message.versions[message.index].content
-							};
-						} else if (message.role === 'assistant') {
-							return {
-								role: message.role,
-								content: message.versions[message.index].content
-							};
-						}
-						return {
-							role: message.role,
-							content: message.versions[message.index].content
-						};
-					}),
+					messages: finalMessages,
 					maxTokens: typeof chat.state.options.maxTokens === 'number' ? chat.state.options.maxTokens : undefined,
-					randomSeed: typeof chat.state.options.randomSeed === 'number' ? chat.state.options.randomSeed : undefined,
+					randomSeed: typeof chat.state.options.seed === 'number' ? chat.state.options.seed : undefined,
 					responseFormat: chat.state.options.json ? { type: 'json_object' } : undefined,
 					safePrompt: chat.state.options.safePrompt,
 					temperature: chat.state.options.temperature,
-					topP: chat.state.options.topP
+					topP: chat.state.options.topP,
+					frequencyPenalty: chat.state.options.frequencyPenalty,
+					presencePenalty: chat.state.options.presencePenalty
 				},
 				{ fetchOptions: { signal: abortController.signal } }
 			);
@@ -204,21 +201,12 @@
 		const outputNode = document.getElementById('messages-container');
 		event.preventDefault();
 		error = null;
-		if (systemPrompt) {
-			chat.state.messages.push({
-				id: uuid(),
-				index: 0,
-				role: 'system',
-				versions: [{ content: [{ type: 'text', text: systemPrompt }] }]
-			});
-			scrollDown(outputNode);
-		}
-		if (promptText.length) {
+		if (promptText.length || files.length) {
 			const message: Message = {
 				id: uuid(),
 				index: 0,
 				role: 'user',
-				versions: [{ content: [{ type: 'text', text: promptText }] }]
+				versions: [{ content: promptText.length ? [{ type: 'text', text: promptText }] : [] }]
 			};
 			if (files.length) {
 				for (let index = 0; index < files.length; index++) {
@@ -255,20 +243,6 @@
 		scrollDown(outputNode);
 
 		await generate(messagesToSend, answer);
-	}
-
-	async function addSystemPrompt(event: Event) {
-		event.preventDefault();
-		if (systemPrompt) {
-			chat.state.messages.push({
-				id: uuid(),
-				index: 0,
-				role: 'system',
-				versions: [{ content: [{ type: 'text', text: systemPrompt }] }]
-			});
-			systemPrompt = '';
-			showOptions = false;
-		}
 	}
 
 	async function stopGenerating(event: Event) {
@@ -375,21 +349,11 @@
 
 	// * < Message events
 
-	function openShare() {
-		const modalComponent: ModalComponent = { ref: ShareModal };
-		const settingsModal: ModalSettings = {
-			type: 'component',
-			backdropClasses: 'bg-gradient-to-tr from-surface-800/50 via-primary-800/50 to-secondary-800/50',
-			component: modalComponent
-		};
-		modalStore.trigger(settingsModal);
-	}
-
 	function handleUploadedFiles(uploadedFiles: File[]) {
 		const { files: validFiles, errors } = handleFileUpload(uploadedFiles);
 		files.push(...validFiles);
 		for (const error of errors) {
-			toastStore.trigger({ message: error, background: 'variant-filled-warning' });
+			toast.error(error);
 		}
 	}
 
@@ -477,279 +441,156 @@
 	});
 </script>
 
-<div
-	class="flex flex-grow flex-shrink justify-center items-stretch flex-col gap-4 p-4 max-h-[calc(100vh-88px)] lg:max-h-screen"
->
-	<div class="relative flex-grow flex-shrink w-full overflow-auto">
-		{#if chat.state.messages.length}
-			<Messages
-				messages={chat.state.messages}
-				interact={{
-					moveUp,
-					moveDown,
-					refresh,
-					previousVersion,
-					nextVersion,
-					deleteVersion,
-					updateMessage,
-					deleteMessage,
-					generate: onSubmit
-				}}
-				{loading}
-				{error}
-			/>
-		{:else}
-			<div class="flex justify-center items-center flex-grow flex-shrink w-full overflow-auto"></div>
-		{/if}
-		<div class="h-4 bg-gradient-to-b from-surface-900/0 to-surface-900 sticky bottom-0 left-0 right-0"></div>
+<div class="flex max-h-[calc(100vh-80px)] shrink grow flex-row gap-0">
+	<div class="hidden lg:flex">
+		<Options />
 	</div>
-	<form class="flex flex-col gap-2 flex-shrink-0" use:focusTrap={true} onsubmit={onSubmit}>
-		{#if !stateIsValid}
-			<aside class="alert variant-ghost-error" transition:slide={{ axis: 'y' }}>
-				<div class="alert-message">
-					<p>
-						Your input is invalid, the first message can only be a system prompt or a question and the last message must
-						be a question or system prompt.
-					</p>
-				</div>
-			</aside>
-		{/if}
-		<ModelError />
-		<label class="label">
-			<div class="flex justify-between items-center gap-2">
-				<div class="flex items-center gap-2">
-					{#if chat.state.options.model}
-						<div class="flex items-center gap-2 text-xs opacity-75 text-right text-primary-500">
-							<span class="badge variant-soft-secondary">Model</span>
-							<div>{chat.state.options.model}</div>
-						</div>
-					{:else}
-						<span></span>
-					{/if}
-					{#if chat.state.usage}
-						<div class="flex items-center gap-2 text-xs opacity-75 text-right text-primary-500">
-							<span class="badge variant-soft-secondary">Tokens</span>
-							<div>
-								Prompt: <span class="text-primary-400">{chat.state.usage.promptTokens}</span> / Completion:
-								<span class="text-primary-400">{chat.state.usage.completionTokens}</span>
-								/ Total: <span class="text-primary-400">{chat.state.usage.totalTokens}</span>
-							</div>
-							{#if chat.state.usage.tps}
-								<div>
-									<span class="text-primary-400">(</span>{chat.state.usage.tps}
-									<span class="text-primary-400">tk/s</span><span class="text-primary-400">)</span>
-								</div>
-							{/if}
-							<span></span>
-						</div>
-					{:else}
-						<span></span>
-					{/if}
-					{#if chat.state.messages.length}
-						<button type="button" class="btn text-xs py-1 px-2 variant-ghost-secondary" onclick={() => openShare()}>
-							Share
-						</button>
-					{/if}
-				</div>
-				{#if tokens > 0}
-					<span class="text-xs" transition:fade>
-						~<span class="text-surface-300">{tokens}</span> tokens
-					</span>
-				{/if}
-			</div>
-			<div class="relative">
-				<textarea
-					bind:value={promptText}
-					disabled={loading || !!models.error}
-					class="textarea"
-					rows="5"
-					placeholder="Type something or drag and drop images..."
-					data-focusindex="0"
-				></textarea>
-				{#if isDragging}
-					<div
-						class="absolute dropzone textarea flex flex-row gap-2 items-center border-2 border-dashed !border-primary-500 p-4 py-4 rounded-container-token top-0 left-0 right-0 bottom-0"
-					>
-						<div class="flex-grow-0 flex-shrink-0">
-							<FileTextIcon class="mx-auto" size="32" />
-						</div>
-						<div class="flex flex-col gap-1 flex-grow flex-shrink">
-							<div class="flex flex-col gap-1 flex-grow flex-shrink">
-								<span class="label-text"><strong>Drop a file</strong></span>
-							</div>
-							<span>Images (.png, .jpeg, .jpg and .webp) allowed.</span>
-						</div>
-					</div>
-				{/if}
-			</div>
-		</label>
-		{#if files.length}
-			<div class="flex flex-col gap-1">
-				{#each files as file, index}
-					<FileUploadPreview {file} {loading} remove={() => files.splice(index, 1)} />
-				{/each}
-			</div>
-		{/if}
-		<div class="flex flex-row justify-between">
-			<button
-				class="btn variant-ghost-surface"
-				type="button"
-				disabled={loading || !!models.error}
-				onclick={(event) => {
-					event.preventDefault();
-					return (showOptions = !showOptions);
-				}}
-			>
-				<Settings2Icon size={20} />
-				<span class="hidden md:inline-block">Options</span>
-			</button>
-			<label for="fileUpload">
-				<input
-					id="fileUpload"
-					class="hidden"
-					type="file"
-					multiple
-					accept="image/png,image/jpeg,image/jpg,image/webp"
-					disabled={loading || models.loading || !!models.error}
-					onchange={onFileChange}
+	<div class="relative flex h-full w-full shrink grow flex-col gap-4">
+		<div class="flex-1 overflow-y-auto px-4">
+			{#if chat.state.messages.length}
+				<Messages
+					messages={chat.state.messages}
+					interact={{
+						moveUp,
+						moveDown,
+						refresh,
+						previousVersion,
+						nextVersion,
+						deleteVersion,
+						updateMessage,
+						deleteMessage,
+						generate: onSubmit
+					}}
+					{loading}
+					{error}
 				/>
-				<button
-					type="button"
-					class="btn variant-filled-secondary transition-all"
-					disabled={loading || models.loading || !!models.error}
-					onclick={() => document.getElementById('fileUpload')?.click()}
-				>
-					<ImageUpIcon size={20} class="md:hidden" />
-					<span class="hidden md:inline-block">Upload image</span>
-				</button>
-			</label>
-			<div class="flex flex-row gap-2">
-				{#if loading}
-					<button class="block btn variant-ghost-error transition-all" type="button" onclick={stopGenerating}>
-						Stop
-					</button>
-				{:else}
-					<button
-						type="submit"
-						class="btn variant-filled-primary transition-all"
-						disabled={loading ||
-							models.loading ||
-							!!models.error ||
-							(!promptText &&
-								(chat.state.messages.length === 0 ||
-									(chat.state.messages.length > 0 &&
-										chat.state.messages[chat.state.messages.length - 1].role !== 'user')))}
-					>
-						Submit
-					</button>
-				{/if}
-			</div>
+			{:else}
+				<div class="flex h-full w-full items-center justify-center"></div>
+			{/if}
 		</div>
-		{#if showOptions}
-			<div class="flex flex-col gap-2" transition:slide={{ axis: 'y' }}>
-				<div class="grid grid-cols-3 gap-2">
-					<div class="flex items-center gap-1">
-						<select bind:value={chat.state.options.model} class="select flex-grow-0" disabled={models.loading}>
-							{#each Object.entries(files.length || chat.state.messages.some( (m) => m.versions.some( (v) => v.content.some((c) => c.type === 'image_url' || c.type === 'document_url') ) ) ? models.visionGroups : models.chatGroups) as [groupName, items]}
-								<optgroup label={groupName}>
-									{#each items as item}
-										<option value={item.id}>{item.id}</option>
-									{/each}
-								</optgroup>
-							{/each}
-						</select>
-						<a href="https://docs.mistral.ai/guides/model-selection/" target="_blank" rel="noreferrer noopener">
-							<CircleHelpIcon />
-						</a>
+		<form class="flex shrink-0 flex-col gap-2 px-4" onsubmit={onSubmit}>
+			{#if !stateIsValid}
+				<aside class="alert variant-ghost-error" transition:slide={{ axis: 'y' }}>
+					<div class="alert-message">
+						<p>
+							Your input is invalid, the first message can only be a system prompt or a question and the last message
+							must be a question or system prompt.
+						</p>
 					</div>
-					<label class="flex-shrink-0">
-						<div class="flex flex-row justify-between items-center">
-							<span>Temperature</span>
-							<span class="text-surface-300">{chat.state.options.temperature}</span>
+				</aside>
+			{/if}
+			<ModelError />
+			<label class="flex flex-col gap-1.5">
+				<div class="flex items-center justify-between gap-2">
+					<div class="flex items-center gap-2">
+						{#if chat.state.usage}
+							<div class="text-primary-500 flex items-center gap-2 text-right text-xs opacity-75">
+								<Badge>Tokens</Badge>
+								<div>
+									Prompt: <span class="text-stone-400">{chat.state.usage.promptTokens}</span> / Completion:
+									<span class="text-stone-400">{chat.state.usage.completionTokens}</span>
+									/ Total: <span class="text-stone-400">{chat.state.usage.totalTokens}</span>
+								</div>
+								{#if chat.state.usage.tps}
+									<div>
+										<span class="text-stone-400">(</span>{chat.state.usage.tps}
+										<span class="text-stone-400">tk/s</span><span class="text-stone-400">)</span>
+									</div>
+								{/if}
+								<span></span>
+							</div>
+						{:else}
+							<span></span>
+						{/if}
+					</div>
+					{#if tokens > 0}
+						<span class="text-xs" transition:fade>
+							~<span class="text-stone-300">{tokens}</span> tokens
+						</span>
+					{/if}
+				</div>
+				<div class="relative">
+					<Textarea
+						rows={5}
+						disabled={loading || !!models.error}
+						placeholder="Type something or drag and drop images..."
+						bind:value={promptText}
+					/>
+					{#if isDragging}
+						<div
+							class="dropzone textarea !border-primary-500 rounded-container-token absolute top-0 right-0 bottom-0 left-0 flex flex-row items-center gap-2 border-2 border-dashed p-4 py-4"
+						>
+							<div class="shrink-0 grow-0">
+								<FileTextIcon class="mx-auto" size="32" />
+							</div>
+							<div class="flex shrink grow flex-col gap-1">
+								<div class="flex shrink grow flex-col gap-1">
+									<span class="label-text"><strong>Drop a file</strong></span>
+								</div>
+								<span>Images (.png, .jpeg, .jpg and .webp) allowed.</span>
+							</div>
 						</div>
+					{/if}
+				</div>
+			</label>
+			{#if files.length}
+				<div class="grid grid-cols-2 gap-4 lg:grid-cols-5">
+					{#each files as file, index}
+						<FileUploadPreview {file} {loading} remove={() => files.splice(index, 1)} />
+					{/each}
+				</div>
+			{/if}
+			<div class="flex flex-row justify-between">
+				<div class="flex flex-row items-center gap-2">
+					<Drawer.Root direction="right">
+						<Drawer.Trigger class="block lg:hidden" type="button" onclick={(event) => event.stopImmediatePropagation()}>
+							<SlidersHorizontalIcon size={20} />
+						</Drawer.Trigger>
+						<Drawer.Content class="flex max-h-screen overflow-auto p-4">
+							<Options />
+						</Drawer.Content>
+					</Drawer.Root>
+					<label for="fileUpload">
 						<input
-							bind:value={chat.state.options.temperature}
-							type="range"
-							name="temperature"
-							id="temperature"
-							min="0"
-							max="1"
-							step="0.01"
-							placeholder="Temperature"
+							id="fileUpload"
+							class="hidden"
+							type="file"
+							multiple
+							accept="image/png,image/jpeg,image/jpg,image/webp"
+							disabled={loading || models.loading || !!models.error}
+							onchange={onFileChange}
 						/>
-					</label>
-					<label class="flex-shrink-0">
-						<div class="flex flex-row justify-between items-center">
-							<span>Top P</span>
-							<span class="text-surface-300">{chat.state.options.topP}</span>
-						</div>
-						<input
-							bind:value={chat.state.options.topP}
-							type="range"
-							name="topP"
-							id="topP"
-							min="0"
-							max="1"
-							step="0.01"
-							placeholder="Top P"
-						/>
+						<Button
+							variant="secondary"
+							disabled={loading || models.loading || !!models.error}
+							onclick={() => document.getElementById('fileUpload')?.click()}
+						>
+							<ImageUpIcon size={20} />
+							<span class="hidden md:inline-block">Upload image</span>
+						</Button>
 					</label>
 				</div>
-				<div class="grid grid-cols-2 lg:grid-cols-4 gap-2 items-center">
-					<div class="flex-shrink-0 cursor-pointer">
-						<SlideToggle name="json" bind:checked={chat.state.options.json}>JSON</SlideToggle>
-					</div>
-					<input
-						bind:value={chat.state.options.maxTokens}
-						class="input"
-						type="number"
-						name="maxTokens"
-						id="maxTokens"
-						min="1"
-						max={models.chat.find((model) => model.id === chat.state.options.model)?.maxContextLength ?? 32000}
-						placeholder="Max tokens"
-					/>
-					<input
-						bind:value={chat.state.options.randomSeed}
-						class="input"
-						type="number"
-						name="randomSeed"
-						id="randomSeed"
-						placeholder="Seed"
-					/>
-					<div class="flex-shrink-0 cursor-pointer">
-						<SlideToggle name="safePrompt" bind:checked={chat.state.options.safePrompt}>Safe prompt</SlideToggle>
-					</div>
-				</div>
-				<div class="flex flex-col gap-2 col-span-2 lg:col-span-3">
-					<label class="label">
-						<div class="flex justify-between items-center">
-							<span>System prompt</span>
-							{#if systemPromptTokens > 0}
-								<span class="text-xs" transition:fade>
-									~<span class="text-surface-300">{systemPromptTokens}</span> tokens
-								</span>
-							{/if}
-						</div>
-						<textarea
-							bind:value={systemPrompt}
-							class="textarea"
-							name="system"
-							id="system"
-							placeholder="System prompt"
-							rows="5"
-						></textarea>
-					</label>
-					<button
-						type="button"
-						class="btn variant-filled-primary transition-all ml-auto"
-						disabled={loading || !systemPrompt}
-						onclick={addSystemPrompt}
-					>
-						Add system prompt
-					</button>
+				<div class="flex flex-row gap-2">
+					{#if loading}
+						<Button variant="destructive" onclick={stopGenerating}>Stop</Button>
+					{:else}
+						<Button
+							type="submit"
+							disabled={loading ||
+								models.loading ||
+								!!models.error ||
+								(!promptText &&
+									(chat.state.messages.length === 0 ||
+										(chat.state.messages.length > 0 &&
+											chat.state.messages[chat.state.messages.length - 1].role !== 'user')) &&
+									files.length === 0)}
+						>
+							Submit
+							<SendHorizontalIcon />
+						</Button>
+					{/if}
 				</div>
 			</div>
-		{/if}
-	</form>
+		</form>
+	</div>
 </div>
