@@ -1,31 +1,32 @@
 <script lang="ts">
-	import { slide } from 'svelte/transition';
-	import { focusTrap, FileDropzone, getToastStore } from '@skeletonlabs/skeleton';
 	import { apiKey } from '$lib/stores/apiKey';
 	import { goto } from '$app/navigation';
 	import { browser } from '$app/environment';
 	import { settings } from '$lib/stores/settings';
 	import { onMount } from 'svelte';
 	import { ocr } from '$lib/stores/ocr.svelte';
-	import Settings2Icon from 'lucide-svelte/icons/settings-2';
-	import FileTextIcon from 'lucide-svelte/icons/file-text';
+	import FileTextIcon from '@lucide/svelte/icons/file-text';
 	import { loadModels, models } from '$lib/stores/models.svelte';
 	import { getClientForRequest } from '$lib/mistral';
 	import ModelError from '$lib/components/ModelError.svelte';
 	import { defaultChatModel } from '$lib/const';
-	import PdfPages from '$lib/components/PdfPages.svelte';
+	import PdfPages from '$lib/components/OCR/PdfPages.svelte';
 	import prettyBytes from 'pretty-bytes';
 	import { fileToB64 } from '$lib/files';
 	import { db } from '$lib/stores/db';
+	import { toast } from 'svelte-sonner';
+	import { FileDropZone, MEGABYTE, type FileDropZoneProps } from '$lib/components/ui/file-drop-zone';
+	import Button from '$lib/components/ui/button/button.svelte';
+	import Badge from '$lib/components/ui/badge/badge.svelte';
+	import Options from './Options.svelte';
+	import * as Drawer from '$lib/components/ui/drawer/index.js';
+	import SlidersHorizontalIcon from '@lucide/svelte/icons/sliders-horizontal';
 
 	if (browser && !$apiKey) {
 		goto('/', { replaceState: true });
 	}
 
-	const toastStore = getToastStore();
-
-	let files = $state<FileList | undefined>(undefined);
-	let showOptions = $state(false);
+	let files = $state<File[] | undefined>(undefined);
 	let error: { text: string; body?: object } | null = $state(null);
 
 	$effect(() => {
@@ -51,18 +52,16 @@
 		ocr.reset();
 
 		if (file.size > 50 * 1024 * 1024) {
-			toastStore.trigger({ message: 'File size should be less than 50MB.' });
+			toast.info('File size should be less than 50MB.');
 			return;
 		}
 
 		const outputNode = document.getElementById('pages-container');
 		loading = true;
-		showOptions = false;
 
 		const b64File = await fileToB64(file);
 
 		abortController = new AbortController();
-		const startedAt = performance.now();
 		try {
 			const client = getClientForRequest({ apiKey: $apiKey, endpoint: $settings.endpoint });
 			const ocrResponse = await client.ocr.process(
@@ -71,7 +70,9 @@
 					document: file.type.includes('image')
 						? { imageUrl: b64File, type: 'image_url' }
 						: { documentUrl: b64File, type: 'document_url' },
-					includeImageBase64: true
+					includeImageBase64: true,
+					imageMinSize: ocr.state.options.minSize,
+					imageLimit: ocr.state.options.imageLimit
 				},
 				{ fetchOptions: { signal: abortController.signal } }
 			);
@@ -104,6 +105,10 @@
 		}
 	}
 
+	const onUpload: FileDropZoneProps['onUpload'] = async (uploadedFiles) => {
+		files = uploadedFiles;
+	};
+
 	async function onSubmit(event: Event) {
 		if (files?.length) {
 			const outputNode = document.getElementById('pages-container');
@@ -126,121 +131,80 @@
 	});
 </script>
 
-<div
-	class="flex flex-grow flex-shrink justify-center items-stretch flex-col gap-4 p-4 max-h-[calc(100vh-88px)] lg:max-h-screen"
->
-	{#if ocr.state.pages.length}
-		<PdfPages pages={ocr.state.pages} {loading} {error} />
-	{:else}
-		<div class="flex justify-center items-center flex-grow flex-shrink w-full overflow-auto"></div>
-	{/if}
-	<form class="flex flex-col gap-2 flex-shrink-0" use:focusTrap={true} onsubmit={onSubmit}>
-		<ModelError />
-		<label class="label">
-			<div class="flex justify-between items-center gap-2">
-				<div class="flex items-center gap-2 truncate">
-					{#if ocr.state.options.model}
-						<div class="flex items-center gap-2 text-xs opacity-75 text-right text-primary-500">
-							<span class="badge variant-soft-secondary">Model</span>
-							<div>{ocr.state.options.model}</div>
-						</div>
-					{/if}
-					{#if ocr.state.usage}
-						<div class="flex items-center gap-2 text-xs opacity-75 text-right text-primary-500">
-							<span class="badge variant-soft-secondary">Document</span>
-							<div>
-								Pages: <span class="text-primary-400">{ocr.state.usage.pagesProcessed}</span>
-							</div>
-						</div>
-					{/if}
-					{#if ocr.state.filename}
-						<div class="text-xs text-primary-500 truncate" title={ocr.state.filename}>{ocr.state.filename}</div>
-					{/if}
-				</div>
-			</div>
-			{#if !files?.length}
-				<FileDropzone
-					bind:files
-					name="files"
-					multiple={false}
-					accept="application/pdf,image/png,image/jpeg,image/jpg,image/webp"
-				>
-					<svelte:fragment slot="lead">
-						<FileTextIcon class="mx-auto" size="32" />
-					</svelte:fragment>
-					<svelte:fragment slot="message">
-						<span class="label-text"><strong>Upload a file</strong> or drag and drop</span>
-					</svelte:fragment>
-					<svelte:fragment slot="meta">
-						<span>PDF and images (.png, .jpeg, .jpg and .webp) allowed.</span>
-					</svelte:fragment>
-				</FileDropzone>
+<div class="flex max-h-[calc(100vh-80px)] shrink grow flex-row gap-0">
+	<div class="hidden lg:flex">
+		<Options />
+	</div>
+	<div class="relative flex h-full w-full shrink grow flex-col gap-4">
+		<div class="flex-1 overflow-y-auto px-4">
+			{#if ocr.state.pages.length}
+				<PdfPages pages={ocr.state.pages} {loading} {error} />
 			{:else}
-				{@const file = files[0]}
-				<div
-					class="dropzone textarea relative flex flex-row gap-2 items-center border-2 border-dashed !border-primary-500 p-4 py-4 rounded-container-token"
-				>
-					<div class="flex-grow-0 flex-shrink-0">
-						<FileTextIcon size="48" />
-					</div>
-					<div class="flex flex-col gap-1 flex-grow flex-shrink">
-						<div class="font-bold">{file.name}</div>
-						<div>{prettyBytes(file.size)}</div>
-					</div>
-					<div>
-						<button type="reset" class="btn variant-ghost-error" disabled={loading} onclick={() => (files = undefined)}>
-							Remove file
-						</button>
+				<div class="flex w-full shrink grow items-center justify-center overflow-auto"></div>
+			{/if}
+		</div>
+		<form class="flex shrink-0 flex-col gap-2" onsubmit={onSubmit}>
+			<ModelError />
+			<label class="flex flex-col gap-1.5">
+				<div class="flex items-center justify-between gap-2">
+					<div class="flex items-center gap-2 truncate">
+						{#if ocr.state.usage}
+							<div class="text-primary-500 flex items-center gap-2 text-right text-xs opacity-75">
+								<Badge>Document</Badge>
+								<div>
+									Pages: <span class="text-primary-400">{ocr.state.usage.pagesProcessed}</span>
+								</div>
+							</div>
+						{/if}
+						{#if ocr.state.filename}
+							<div class="text-primary-500 truncate text-xs" title={ocr.state.filename}>{ocr.state.filename}</div>
+						{/if}
 					</div>
 				</div>
-			{/if}
-		</label>
-		<div class="flex flex-row justify-between">
-			<button
-				class="btn variant-ghost-surface"
-				type="button"
-				disabled={loading || !!models.error}
-				onclick={(event) => {
-					event.preventDefault();
-					return (showOptions = !showOptions);
-				}}
-			>
-				<Settings2Icon size={20} />
-				<span>Options</span>
-			</button>
-			<div class="flex flex-row gap-2">
-				<!-- {#if ocr.state.messages.length}
-					<button type="button" class="btn variant-ghost-secondary mx-auto" onclick={() => openShare()}>Share</button>
-				{/if} -->
-				{#if loading}
-					<button class="btn variant-ghost-error transition-all" type="button" onclick={stopGenerating}> Stop </button>
+				{#if !files?.length}
+					<FileDropZone
+						{onUpload}
+						name="files"
+						maxFiles={1}
+						maxFileSize={10 * MEGABYTE}
+						accept="application/pdf,image/png,image/jpeg,image/jpg,image/webp"
+						fileCount={files?.length ?? 0}
+					/>
 				{:else}
-					<button
-						type="submit"
-						class="btn variant-filled-primary transition-all"
-						disabled={loading || models.loading || !!models.error || !files?.length}
+					{@const file = files[0]}
+					<div
+						class="border-border hover:bg-accent/25 flex h-48 w-full flex-row place-items-center items-center justify-center gap-2 rounded-lg border-2 border-dashed p-6 transition-all"
 					>
-						Submit
-					</button>
+						<div class="shrink-0 grow-0">
+							<FileTextIcon size="48" />
+						</div>
+						<div class="flex shrink grow flex-col gap-1">
+							<div class="font-bold">{file.name}</div>
+							<div>{prettyBytes(file.size)}</div>
+						</div>
+						<div>
+							<Button type="reset" variant="destructive" disabled={loading} onclick={() => (files = undefined)}>
+								Remove file
+							</Button>
+						</div>
+					</div>
+				{/if}
+			</label>
+			<div class="flex flex-row justify-between gap-2 lg:justify-end">
+				<Drawer.Root direction="right">
+					<Drawer.Trigger class="block lg:hidden" type="button" onclick={(event) => event.stopImmediatePropagation()}>
+						<SlidersHorizontalIcon size={20} />
+					</Drawer.Trigger>
+					<Drawer.Content class="flex max-h-screen overflow-auto p-4">
+						<Options />
+					</Drawer.Content>
+				</Drawer.Root>
+				{#if loading}
+					<Button variant="destructive" type="button" onclick={stopGenerating}>Stop</Button>
+				{:else}
+					<Button type="submit" disabled={loading || models.loading || !!models.error || !files?.length}>Submit</Button>
 				{/if}
 			</div>
-		</div>
-		{#if showOptions}
-			<div class="flex flex-col gap-2" transition:slide={{ axis: 'y' }}>
-				<div class="grid grid-cols-3 gap-2">
-					<div class="flex items-center gap-1">
-						<select bind:value={ocr.state.options.model} class="select flex-grow-0" disabled={models.loading}>
-							{#each Object.entries(models.ocrGroups) as [groupName, items]}
-								<optgroup label={groupName}>
-									{#each items as item}
-										<option value={item.id}>{item.id}</option>
-									{/each}
-								</optgroup>
-							{/each}
-						</select>
-					</div>
-				</div>
-			</div>
-		{/if}
-	</form>
+		</form>
+	</div>
 </div>
