@@ -1,14 +1,10 @@
+import tailwindcss from '@tailwindcss/vite';
 import wasm from 'vite-plugin-wasm';
 import { sveltekit } from '@sveltejs/kit/vite';
 import { enhancedImages } from '@sveltejs/enhanced-img';
-import { defineConfig } from 'vitest/config';
+import { defineConfig } from 'vite';
 
 export default defineConfig({
-	plugins: [wasm(), enhancedImages(), sveltekit()],
-	build: {
-		target: 'esnext'
-	},
-	test: {
-		include: ['src/**/*.{test,spec}.{js,ts}']
-	}
+	plugins: [tailwindcss(), wasm(), enhancedImages(), sveltekit()],
+	build: { target: 'esnext' }
 });
